+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [v0.0.?] - 2020/09/09
+
+* Adds ability to accept an Earthdata Login access token and make download requests
+  using that token, falling back to the application credentials if it isn't provided.
+
 ## [v0.0.10] - 2020/06/18
 
 * Updates the install instructions in the README that use the Earthdata Maven repository